encryptions/python/requirements.txt
## The following requirements were added by pip freeze:
alembic==1.11.1
async-timeout==4.0.2
cffi==1.15.1
cryptography==41.0.1
dnspython==2.3.0
ed25519==1.5
Faker==18.11.2
greenlet==2.0.2
hiredis==2.2.3
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.25.0
pandas==2.0.3
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pymongo==4.4.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
tzdata==2023.3
